requirements.txt
Django==4.2.7
django-admin==2.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
